requirements.txt
python-telegram-bot==20.5
requests==2.31.0
python-dotenv==1.0.1
ai21==2.1.4  # Official AI21 SDK